Dutch Auction Attempt
Most changes are quite self-explanatory... The dutch auction requires a
timer so that the auction progresses if no-one wants to respond...
Doesnt seem to function yet. Also added some code to allow for minimal
prices in AuctionItems for the dutch auction. Based most of the code on
the Behaviours for the Second-Price auction